requirements.txt
numpy
pandas
torch
torchvision
matplotlib
pillow
tqdm